clean up and added msg_sender

+++ examples/voting/src/message.rs
@@ -5,6 +5,6 @@ pub enum BallotMsg {
     #[n(0)]
     Vote {
         #[n(0)]
-        p: usize,
+        p: u32,
     },
 }
